==> 03-AFDM_AG.Rmd <==
---
title: "AFDM"
author: "Marine Boudy"
date: '2022-04-28'
output: html_document
---

```{r librairies_3}

library(tidyverse)
library(knitr)
library(ggplot2)
library(emmeans)
library(cowplot)
library(corrplot)
library(factoextra)
library(FactoMineR)
library(ppcor)
library(multcomp)

```

# AFDM:Analyse factorielle de données mixtes
## Angélique

```{r}
load("vault/data/AG_plac.RData")
```

#1er jeu de donnees

```{r}
AG_plac_AFDM <- AG_plac %>%
  dplyr::select(Foret,Type.placette,diversite,n_AG,n_inf_130,n_sup_130, hauteur_moy, hauteur_min, hauteur_max,sumhauteur) %>% 
  filter(!is.na(hauteur_min),!is.na(hauteur_moy),!is.na(hauteur_max), !is.na(sumhauteur),!is.na(Type.placette))
 # dplyr::select(Foret,n:n_inf_130,n_sup_130,diversite:n_VIO)
sum(is.na(AG_plac_AFDM))
  
  
  # dplyr::select(2:10,16:22,26,27,31,34,37,40,41,44,45,48,49,51,52,56,57,60,61)

res.famd <- FAMD(AG_plac_AFDM, graph = FALSE)

```

```{r valeurs_propres}
eig.val <- res.famd$eig
barplot(eig.val[, 2], 
        names.arg = 1:nrow(eig.val), 
        main = "Variances Explained by Dimensions (%)",
        xlab = "Principal Dimensions",
        ylab = "Percentage of variances",
        col ="steelblue")
# Add connected line segments to the plot
lines(x = 1:nrow(eig.val), eig.val[, 2], 
      type = "b", pch = 19, col = "red")

eig.val
```
## Variables qualitatives et quantitatives

```{r}
var <- get_famd_var (res.famd)
var

# Coordonnées des variables
head(var$coord)
# Cos2: qualité de représentation
head(var$cos2)
# Contributions aux dimensions
head(var$contrib)
```

```{r}
# Graphique des variables
fviz_famd_var (res.famd, repel = TRUE)
fviz_famd_var (res.famd, repel = TRUE)
# Contribution à la première dimension
fviz_contrib (res.famd, "var", axes = 1)
# Contribution à la deuxième dimension
fviz_contrib (res.famd, "var", axes = 2)
# Contribution à la 3e dimension
fviz_contrib (res.famd, "var", axes = 3)
```

## Variables quantitatives

```{r}
quanti.var <- get_famd_var(res.famd, "quanti.var")
quanti.var 

fviz_famd_var(res.famd, "quanti.var",
              col.var = "black")# on ne peut pas utiliser repel=TRUE car trop de variables
#fviz_famd_var(res.famd, "quanti.var", repel = TRUE,col.var = "black")

#couleur selon importance de la contribution aux axes
fviz_famd_var(res.famd, "quanti.var", col.var = "contrib", 
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"))

# Couleur par valeurs cos2: qualité sur le plan des facteurs
fviz_famd_var(res.famd, "quanti.var", col.var = "cos2",
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"))

fviz_famd_var(res.famd, "quanti.var", col.var = "cos2",
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"), axes = c(1,3))

# Cos2 total des variables sur Dim.1 et Dim.2
fviz_cos2(res.famd, choice = "var", axes = 1:2)

# Cos2 total des variables sur Dim.1 et Dim.3
fviz_cos2(res.famd, choice = "var", axes = 1:3)



```

## Variables qualitatives

```{r}
quali.var <- get_famd_var(res.famd, "quali.var")
quali.var 

fviz_famd_var(res.famd, "quali.var", col.var = "cos2", 
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07")
             )

fviz_famd_var(res.famd, "quali.var", col.var = "cos2", 
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"), axes=c(1,3))
```

## Individus

```{r}
ind <- get_famd_ind(res.famd)
ind

#couleur par contribution aux axes
fviz_famd_ind(res.famd, col.ind = "contrib",
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"))
#couleur par qualité de representation
fviz_famd_ind(res.famd, col.ind = "cos2",
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"))
#couleur par traitement
fviz_mfa_ind(res.famd, 
             habillage = "Foret", # color by groups 
             palette = c("#00AFBB", "#E7B800", "#FC4E07"),
             addEllipses = TRUE, ellipse.type = "confidence", 
             repel = TRUE # Avoid text overlapping
             ) 

fviz_ellipses(res.famd, c("Foret"))


# n'affiche pas le num des individus
fviz_ellipses(res.famd, 1:2, geom = "point")

# individu extreme 

```
