==> 04-paracou_complements.Rmd <==
---
title: "Surfaces d'ouverture"
author: "Marine Boudy"
date: '2022-05-30'
output:
  pdf_document: default
  html_document: default
---

# Surfaces d'ouverture de la canopee

L'objectif est de comparer les surfaces de trouees intersectées dans un rayon de 5,10 et 15m autour des palcettes entre Paracou et les autres sites d'études, afin de rééquilibrer échantillonnage.

On etudie les 3 variables suivantes:

P5.S = surface de trouée localisée dans un rayon de 5m autour du centre de la placette d'inventaire
P10.S = surface de trouée localisée dans un rayon de 10m autour du centre de la placette d'inventaire
P15.S = surface de trouée localisée dans un rayon de 15m autour du centre de la placette d'inventaire


```{r librairies_4}
library(tidyverse)
library(sf)
library(ggplot2)
library(patchwork)
library(units)


```

```{r donnees, warning=FALSE}

load("vault/data/ouvertures.RData")

ouvertures_histo <-ouvertures %>% 
  drop_units() %>% # les surfaces d'intersection avec les trouee ont une unite(m2) ,on la retire pour le moment(probleme d'affichage avec ggplot2)
  mutate(Site = case_when(Foret=="Paracou" ~ "Paracou",
                           Parcelle=="HKO096" ~ "HKO096",
                           Parcelle=="MAW745" ~ "MAW",
                           Parcelle=="MAW744" ~ "MAW",
                           Parcelle=="PAI74" ~ "PAI74"))# variable "Site d'etude" pour les facet_wrap
ouvertures_histo[is.na(ouvertures_histo)] <- 0


ouvertures_paracou <- ouvertures_histo %>% 
  filter(Foret=="Paracou")#selection des données de Paracou

ouvertures_autres <- ouvertures_histo %>% 
  filter(!Foret=="Paracou")#selecction des données hors Paracou

```


```{r}
####theme plot

theme_mb <- function(base_size = 60) {
  theme_bw(base_size = base_size) %+replace%
    theme(
      # L'ensemble de la figure
      plot.title = element_text(size = rel(1), face = "bold", margin = margin(0,0,5,0), hjust = 0),
      # Zone où se situe le graphique
      panel.grid.minor = element_blank(),
      panel.border = element_blank(),
      # Les axes
      axis.title = element_text(size = rel(0.85), face = "bold"),
      axis.text = element_text(size = rel(0.70), face = "bold"),
      axis.line = element_line(color = "black", arrow = arrow(length = unit(0.3, "lines"), type = "closed")),
      #La légende
      legend.title = element_text(size = rel(0.85), face = "bold"),
      legend.text = element_text(size = rel(0.70), face = "bold"),
      legend.key = element_rect(fill = "transparent", colour = NA),
      legend.key.size = unit(1.5, "lines"),
      legend.background = element_rect(fill = "transparent", colour = NA),
      # Les étiquettes dans le cas d'un facetting
      strip.background = element_rect(fill = "#7AC5CD", color = "#7AC5CD"),
      strip.text = element_text(size = rel(0.85), face = "bold", color = "white", margin = margin(5,0,5,0))
      )
}
```


## Comparaison de paracou aux autres sites

```{r }
#Plot comparant Paracou aux autres forets
####theme plot

p1<-ggplot(data = ouvertures_paracou) +
  geom_histogram(aes(x = P5.S),size=2.5, binwidth=1, boundary=0, closed ="left",colour = 4, fill = "white")+
  ggtitle("Paracou")+
  ylab("Effectif")+
  lims(x = c(-1, 90), y = c(0, 200))+
  theme_mb()

p2<-ggplot(data = ouvertures_autres) +
  geom_histogram(aes(x = P5.S),size=2.5,binwidth=1,boundary=0,closed ="left", colour = 4, fill = "white")+
  ggtitle("Autres sites") +
  ylab("Effectif")+
  lims(x = c(-1, 90), y = c(0, 200))+
  theme_mb()
  


p3 <- ggplot(data = ouvertures_paracou) +
  geom_histogram(aes(x = P10.S),size=2.5,binwidth=5, boundary=0,closed ="left",colour = 4, fill = "white")+
  ggtitle("Paracou") +
  ylab("Effectif")+
  lims(x = c(-1, 250), y = c(0, 140))+
  theme_mb()
  
p5 <- ggplot(data = ouvertures_autres) +
  geom_histogram(aes(x = P10.S),size=2.5,binwidth=5, boundary=0, closed ="left",colour = 4, fill = "white")+
  ggtitle("Autres sites") +
  ylab("Effectif")+
  lims(x = c(-1, 250), y = c(0, 140))+
  theme_mb()


p4 <- ggplot(data = ouvertures_paracou) +
  geom_histogram(aes(x = P15.S),size=2.5,binwidth=5, boundary=0,closed ="left",colour = 4, fill = "white")+
  ggtitle("Paracou") +
  ylab("Effectif")+
  lims(x = c(-1, 500), y = c(-1,70))+
  theme_mb()


p6 <- ggplot(data = ouvertures_autres) +
  geom_histogram(aes(x = P15.S),size=2.5,binwidth=5,boundary=0, closed ="left",colour = 4, fill = "white")+
  ggtitle("Autres sites") +
  ylab("Effectif")+
  lims(x = c(-1, 500), y = c(-1,70))+
  theme_mb() 

#--- Plot regroupant chaque site d'étude

p8<- ggplot(data = ouvertures_histo) +
  geom_histogram(aes(P5.S),size=2.5,binwidth=1,boundary=0, closed ="left", colour = 4, fill = "white")+
  facet_wrap(~ Site,ncol=1) +
  ylab("Effectif")+
  theme_mb()


p9<- ggplot(data = ouvertures_histo) +
  geom_histogram(aes(P10.S),size=2.5,binwidth=5,boundary=0, closed ="left", colour = 4, fill = "white")+
  facet_wrap(~ Site,ncol=1)+
  ylab("Effectif")+
  theme_mb()


p10<- ggplot(data = ouvertures_histo) +
  geom_histogram(aes(P15.S),size=2.5,binwidth=5,boundary=0, closed ="left", colour = 4, fill = "white")+
  facet_wrap(~ Site,ncol=1)+
  ylab("Effectif")+
  theme_mb()


```


## Plots

```{r fig.height=50, fig.width=50}


#--- Plot de Paracou et autres sites mélangés(Regina, St Georges)
# (p1 / p2 | p3 /p5 | p4/p6) +  plot_annotation ( title = "Histogramme des surfaces de trouées situées dans un rayon de 5, 10, et 15 m autour des placettes d'inventaire",subtitle = "Comparaison du site de Paracou aux inventaire sur les autres sites")
# 

(p1/p2)+plot_annotation(title = "Comparaison des surfaces de trouées situées dans un rayon de 5 m autour des placettes d'inventaire entre Paracou \n et les autres sites d'étude",theme = theme_mb() ) 

(p3/p5)+plot_annotation(title = "Comparaison des surfaces de trouées situées dans un rayon de 10 m autour des placettes d'inventaire entre Paracou \n et les autres sites d'étude",theme = theme_mb() ) 

(p4/p6)+plot_annotation(title = "Comparaison des surfaces de trouées situées dans un rayon de 15 m autour des placettes d'inventaire entre Paracou \n et les autres sites d'étude",theme = theme_mb() ) 

```

rq:Les surfaces nulles, dont l'effectif est toujours très grand, écrasent les graphes
La largeur des barres des histogrammes est de 1m2 pour P5.S,et de 5m2 pour P10.S et P15.S

Le graphique ci-dessus montre qu'il existe un manque dans les grandes surfaces pour Paracou, ce qui est cohérent car l'exploitation est ancienne, les trouees sont donc de taille réduite.
La repartition des ouvertures de moins de 100m dans le rayon de 10m est également différente.

## Comparaison des 4 sites

```{r fig.height=50, fig.width=50}
# p8 + p9 + p10 +plot_annotation(
#   title = "Histogramme des surfaces de trouées situées dans un rayon de 5, 10, et 15 m \n autour des placettes d'inventaire pour chaque site d'etude")

p8+plot_annotation( title = "Histogramme des surfaces de trouées situées dans un rayon de 5 m \n  autour des placettes d'inventaire pour chaque site d'etude",theme= theme_mb())

p9+plot_annotation(title = "Histogramme des surfaces de trouées situées dans un rayon de 10 m \n autour des placettes d'inventaire pour chaque site d'etude",theme= theme_mb())

p10+plot_annotation(title = "Histogramme des surfaces de trouées situées dans un rayon de 15 m \n autour des placettes d'inventaire pour chaque site d'etude",theme= theme_mb())


```
